+++ src/main/java/com/luoyu/blog/controller/video/VideoController.java
@@ -2,6 +2,8 @@ package com.luoyu.blog.controller.video;
 
 import com.luoyu.blog.common.aop.annotation.LogView;
 import com.luoyu.blog.common.constants.RedisCacheNames;
+import com.luoyu.blog.common.enums.ResponseEnums;
+import com.luoyu.blog.common.exception.MyException;
 import com.luoyu.blog.common.util.PageUtils;
 import com.luoyu.blog.common.validator.ValidatorUtils;
 import com.luoyu.blog.common.validator.group.AddGroup;
@@ -98,6 +100,14 @@ public class VideoController {
     @RequiresPermissions("video:delete")
     @CacheEvict(allEntries = true)
     public Response deleteVideos(@RequestBody Integer[] ids) {
+        if (ids == null || ids.length < 1){
+            throw new MyException(ResponseEnums.PARAM_ERROR.getCode(), "ids不能为空");
+        }
+
+        if (ids.length > 100){
+            throw new MyException(ResponseEnums.PARAM_ERROR.getCode(), "ids不能超过100个");
+        }
+
         videoService.deleteVideos(ids);
         return Response.success();
     }
